Add failed state for packaging and add SystemTools for searching for MSBuild + update default template

+++ GloryEngine/Editor/Extensions/JoltEditorExtension/PhysicsSettings.cpp
@@ -193,7 +193,7 @@ namespace Glory::Editor
 		Utils::NodeValueRef collisionMatrix = RootValue()["CollisionMatrix"];
 		if (!collisionMatrix.Exists())
 		{
-			stream->Write(0);
+			stream->Write(size_t(0));
 			return;
 		}
